requirements.txt
# KiteAPI Trading Framework
# Requirements for Kite Connect API integration

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# HTTP Requests
requests>=2.31.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Date/Time Parsing
python-dateutil>=2.8.2

# WebSocket Client
websocket-client>=1.6.0

# YAML Parsing
pyyaml>=6.0.0

# Async Support
httpx>=0.24.0
aiohttp>=3.8.0

# Database - PostgreSQL
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0

# Cache - Redis
redis>=5.0.0
hiredis>=2.2.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Background Tasks
celery[redis]>=5.3.0

# Monitoring
psutil>=5.9.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
flake8>=6.1.0

# Type Stubs
types-redis>=4.6.0
types-passlib>=1.7.0
types-python-jose>=3.3.0
types-psutil>=5.9.0
cryptography>=42.0.0
Package            Version
------------------ -----------
aiohappyeyeballs   2.6.1
aiohttp            3.13.2
aiosignal          1.4.0
amqp               5.3.1
annotated-doc      0.0.4
annotated-types    0.7.0
anyio              4.11.0
asyncpg            0.31.0
attrs              25.4.0
autoflake          2.3.1
bcrypt             5.0.0
billiard           4.2.3
black              25.11.0
celery             5.5.3
certifi            2025.11.12
cffi               2.0.0
charset-normalizer 3.4.4
click              8.3.1
click-didyoumean   0.3.1
click-plugins      1.1.1.2
click-repl         0.3.0
coverage           7.12.0
cryptography       46.0.3
dnspython          2.8.0
ecdsa              0.19.1
email-validator    2.3.0
fastapi            0.122.0
flake8             7.3.0
frozenlist         1.8.0
greenlet           3.2.4
h11                0.16.0
httpcore           1.0.9
httptools          0.7.1
httpx              0.28.1
idna               3.11
iniconfig          2.3.0
isort              7.0.0
kombu              5.5.4
mccabe             0.7.0
multidict          6.7.0
mypy               1.18.2
mypy_extensions    1.1.0
outcome            1.3.0.post0
packaging          25.0
passlib            1.7.4
pathspec           0.12.1
pip                25.3
platformdirs       4.5.0
pluggy             1.6.0
prompt_toolkit     3.0.52
propcache          0.4.1
psutil             7.1.3
pyasn1             0.6.1
pycodestyle        2.14.0
pycparser          2.23
pydantic           2.12.5
pydantic_core      2.41.5
pyflakes           3.4.0
Pygments           2.19.2
pyotp              2.9.0
PySocks            1.7.1
pytest             9.0.1
pytest-asyncio     1.3.0
pytest-cov         7.0.0
python-dateutil    2.9.0.post0
python-dotenv      1.2.1
python-jose        3.5.0
pytokens           0.3.0
PyYAML             6.0.3
redis              7.1.0
requests           2.32.5
rsa                4.9.1
selenium           4.38.0
six                1.17.0
sniffio            1.3.1
sortedcontainers   2.4.0
SQLAlchemy         2.0.44
starlette          0.50.0
trio               0.32.0
trio-websocket     0.12.2
typing_extensions  4.15.0
typing-inspection  0.4.2
tzdata             2025.2
urllib3            2.5.0
uvicorn            0.38.0
uvloop             0.22.1
vine               5.1.0
watchfiles         1.1.1
wcwidth            0.2.14
websocket-client   1.9.0
websockets         15.0.1
wsproto            1.3.2
yarl               1.22.0
